+++ src/steps/insert-path/IntoStep.ts
@@ -1,3 +1,5 @@
+import { TableAsterisk } from '@src/TableAsterisk'
+import { getMinOneArray } from '../../util'
 import { Column, Table } from '../../database'
 import { Artifacts, BaseStep } from '../BaseStep'
 import { PrimitiveType } from '../../models'
@@ -6,16 +8,18 @@ import { InsertColumnsAndExpressionsNotEqualError } from '../../errors'
 import { Binder } from '../../binder'
 import { SelectItem, SelectStep } from '../select-path/SelectStep'
 import { SelectItemInfo } from '../../SelectItemInfo'
-import { Default } from '../../singletoneConstants'
+import { Asterisk, Default } from '../../singletoneConstants'
 import { ValuesStep } from './ValuesStep'
 
 export abstract class IntoStep extends BaseStep {
+	public values(value: (PrimitiveType|Binder|Default), ...values: (PrimitiveType|Binder|Default)[]): ValuesStep
 	public values(...values: (PrimitiveType|Binder|Default)[]): ValuesStep {
-		return new ValuesStep(this, values)
+		return new ValuesStep(this, getMinOneArray(values))
 	}
 
+	public values$(value: PrimitiveType, ...values: PrimitiveType[]): ValuesStep
 	public values$(...values: PrimitiveType[]): ValuesStep {
-		return new ValuesStep(this, values.map(it => new Binder(it)))
+		return new ValuesStep(this, getMinOneArray(values.map(it => new Binder(it))))
 	}
 
 	public defaultValues(): DefaultValuesStep {
@@ -24,32 +28,27 @@ export abstract class IntoStep extends BaseStep {
 
 	public select(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
 		// TODO: consider adding DISTINCT and ALL to items without effecting matching number of values
+		// TODO: consider adding ASTERISK and Table Asterisk to items without throwing error if matching column number
 		if (items.length === 0) {
 			throw new Error('Invalid empty SELECT step')
 		}
-		this.throwForInvalidExpressionsNumber(items)
+		this.throwWhenInvalidExpressionsNumber(items)
 		return new SelectStep(this, items)
 	}
 
-	private throwForInvalidExpressionsNumber(items: (SelectItemInfo|SelectItem|PrimitiveType)[]) {
-		// TODO: in case columnCount = 0 we should check number of column in schema
-		const columnsCount = this.getStepArtifacts().columns.size
-		if (columnsCount > 0 && columnsCount !== items.length) {
-			throw new InsertColumnsAndExpressionsNotEqualError(columnsCount, items.length)
-		}
-	}
+	protected abstract throwWhenInvalidExpressionsNumber(items: (SelectItemInfo|SelectItem|PrimitiveType)[]): void
 }
 
 export class IntoTableStep extends IntoStep {
 	constructor(
 		prevStep: BaseStep,
-		private readonly table: Table,
+		public readonly table: Table,
 	) {
 		super(prevStep)
 		this.throwIfTableNotInDb(table)
 		return new Proxy(
 			this,
-			{ apply: (target, thisArg, args) => target.selfCall(...args) },
+			{ apply: (target: this, thisArg, args: Column[]) => target.selfCall(...args) },
 		)
 	}
 
@@ -64,13 +63,27 @@ export class IntoTableStep extends IntoStep {
 	private selfCall(...columns: Column[]): IntoColumnsStep {
 		return new IntoColumnsStep(this, columns)
 	}
+
+	protected throwWhenInvalidExpressionsNumber(items: (SelectItemInfo|SelectItem|PrimitiveType)[]): void {
+		if (items.find(it => it instanceof Asterisk)) {
+			/** Validation can not be done when asterisk is used because it is unknown which table(s) is used */
+			return
+		}
+		const expCount = countExpressions(items)
+
+		const tableColumnCount = this.table.getColumns().length
+		if (expCount !== tableColumnCount) {
+			throw new InsertColumnsAndExpressionsNotEqualError(tableColumnCount, expCount)
+		}
+	}
 }
 
 export class IntoColumnsStep extends IntoStep {
 	override prefixSeparator = ''
+
 	constructor(
 		prevStep: BaseStep,
-		private readonly columns: Column[],
+		public readonly columns: Column[],
 	) {
 		super(prevStep)
 	}
@@ -82,4 +95,30 @@ export class IntoColumnsStep extends IntoStep {
 	getStepArtifacts(): Artifacts {
 		return { tables: new Set(), columns: new Set(this.columns) }
 	}
+
+	protected throwWhenInvalidExpressionsNumber(items: (SelectItemInfo|SelectItem|PrimitiveType)[]): void {
+		if (items.find(it => it instanceof Asterisk)) {
+			/** Validation can not be done when asterisk is used because it is unknown which table(s) is used */
+			return
+		}
+		const expCount = countExpressions(items)
+		const columnsCount = this.getStepArtifacts().columns.size
+		if (columnsCount === 0) {
+			throw new Error('IntoColumnsStep must have at least one column')
+		} else if (expCount !== columnsCount) {
+			throw new InsertColumnsAndExpressionsNotEqualError(columnsCount, expCount)
+		}
+	}
+}
+
+function countExpressions(items: (SelectItemInfo|SelectItem|PrimitiveType)[]): number {
+	let expCount = 0
+	items.forEach(it => {
+		if (it instanceof TableAsterisk) {
+			expCount += it.table.getColumns().length
+		} else {
+			expCount += 1
+		}
+	})
+	return expCount
 }